requirements.txt
astroid==3.2.2
autopep8==2.3.0
blinker==1.8.2
cachelib==0.13.0
click==8.1.7
colorama==0.4.6
dill==0.3.8
Flask==3.0.3
Flask-MQTT==1.2.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
msgspec==0.18.6
numpy==2.0.0
paho-mqtt==1.6.1
pandas==2.2.2
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pylint==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
tomlkit==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3
